Fail fast if slashing protection invalid
